+++ Services/MealDataService/Sources/Data/Remote/Response/MealMenuPerMonthListResponse.swift
@@ -1,0 +1,12 @@
+import Foundation
+
+struct MealMenuListPerMonthResponse: Decodable {
+
+    let meals: [MealMenuPerMonthResponse]
+}
+
+extension MealMenuListPerMonthResponse {
+    func toDomain() -> [MealMenuEntity] {
+        return meals.map { $0.toDomain() }
+    }
+}
