+++ Services/MealDataService/Sources/Data/Remote/Response/MealMenuPerMonthResponse.swift
@@ -1,12 +1,21 @@
 import Foundation
 
-struct MealMenuListPerMonthResponse: Decodable {
-
-    let meals: [MealMenuPerMonthResponse]
+struct MealMenuPerMonthResponse: Decodable {
+    let date: String
+    let breakfast: [String]
+    let lunch: [String]
+    let dinner: [String]
 }
 
-extension MealMenuListPerMonthResponse {
-    func toDomain() -> [MealMenuEntity] {
-        return meals.map { $0.toDomain() }
+extension MealMenuPerMonthResponse {
+    func toDomain() -> MealMenuEntity {
+        return .init(
+            date: date.toDate(format: .fullDate),
+            menu: [
+                .breakfast: breakfast,
+                .lunch: lunch,
+                .dinner: dinner
+            ]
+        )
     }
 }
